Fix four review-pass bugs found during audit verification
- clip_dedup.py: filter_shard inline write site (line ~718) was missing
  fsync+rename; it wrote directly to the final FAISS index path.  Fixed
  with the same _idx_tmp+fsync+os.replace pattern used at the other site.

- precompute_all.py: _quantize_int8 and _quantize_int8_batch were dead code
  after the SIGNAL-6 VAE float32 fix (no call sites remained).  Removed.

- test_ip_adapter_inference.py: _flux_forward_no_ip was an unused import
  after CROSS-4 replaced the end-sum path with _flux_forward_with_ip.
  Removed.  Also corrected the _generate docstring which still described
  the old end-sum accumulation approach.

- dataset.py: make_prefetch_loader docstring updated: vae_latents yield
  type changed from float16 to float32 to match the SIGNAL-6 loader fix.

+++ train/ip_adapter/dataset.py
@@ -328,7 +328,7 @@ def make_prefetch_loader(
       images:      float32 numpy [B, C, H+32, W+32] — padded for GPU crop augmentation
       captions:    list of str, len B (empty string if text dropout)
       text_embeds: float16 numpy [B, seq, 7680] or None (if no Qwen3 cache)
-      vae_latents: float16 numpy [B, 32, H/8, W/8] or None (if no VAE cache)
+      vae_latents: float32 numpy [B, 32, H/8, W/8] or None (if no VAE cache)
       siglip_feats:float16 numpy [B, 729, 1152] or None (if no SigLIP cache)
       bucket_hw:   (H, W) tuple for this batch
     """
